support compiled with clang and remove warining

+++ queue.h
@@ -18,21 +18,22 @@
 #define RECDVB_QUEUE_H
 
 #include <stdint.h>
+#include <stddef.h>
 #include <pthread.h>
 
 #define MAX_READ_SIZE           (188 * 87) /* 188*87=16356 */
 
 typedef struct _BUFSZ {
-	int size;
+	ssize_t size;
 	uint8_t buffer[MAX_READ_SIZE];
 } BUFSZ;
 
 typedef struct _QUEUE_T {
 	unsigned int in;           // 次に入れるインデックス
 	unsigned int out;          // 次に出すインデックス
-	unsigned int size;         // キューのサイズ
-	unsigned int num_avail;    // 満タンになると 0 になる
-	unsigned int num_used;     // 空っぽになると 0 になる
+	size_t size;               // キューのサイズ
+	size_t num_avail;          // 満タンになると 0 になる
+	size_t num_used;           // 空っぽになると 0 になる
 	pthread_mutex_t mutex;
 	pthread_cond_t cond_avail; // データが満タンのときに待つための cond
 	pthread_cond_t cond_used;  // データが空のときに待つための cond
